pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dicom2fhir"
version = "0.1.15"
description = "Convert dicoms to FHIR Bundles containing ImagingStudy, Patient, Device, and Observation resources."
keywords = ["dicom", "fhir", "resources", "python", "hl7", "health IT", "healthcare"]
classifiers = [
  "Intended Audience :: Developers",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.12",
  "Topic :: Scientific/Engineering :: Medical Science Apps.",
  "Typing :: Typed"
]
readme = {file = "README.md", content-type = "text/markdown"}
license-files = ["LICENSE"]
authors = [{name = "Felix Nensa", email = "felix.nensa@uk-essen.de"}]
dependencies = [
  "pydicom>=3.0.1",
  "fhir.resources>=8.0.0",
  "html5lib>=1.1",
  "pandas>=2.3.0",
  "PyYAML>=6.0.2",
  "python-dateutil",
  "pydantic>=2.11.5"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.4.0",
  "psycopg2>=2.9.10"
]

[project.urls]
Repository = "https://github.com/UMEssen/dicom-fhir-converter"

[tool.setuptools.package-data]
"dicom2fhir" = ["resources/terminologies/*.json"]

[tool.setuptools.packages.find]
where = ["."]               # scan the project root
include = ["dicom2fhir*"]